pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Smart Meeting Assistant Backend"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "python-dotenv>=1.0.0",
    "vision-agents[gemini,getstream]>=0.5.1",
]